Add and use the autoformat plugin
This adds an :Autoformat command to run formatters like
gofmt/shfmt/yapf/etc, and my vimrc runs this on save for the sh and go
filetypes for now. The fallbacks for retab, indent, and trailing
whitespace removal are disabled.

As opposed to alternative filetype specific plugins, I like that this
handles the common portions itself and lets you manipulate the
formatters used, whereas plugins like vim-go hardcode everything. The
downside is they aren't going to add errors from gofmt to the quickfix
menu and the like, but I rather like the simplicity of this one.

I originally wanted to avoid using a plugin for formatters at all, but
attempts to do so in a more simplistic fashion run into problems, i.e.
use of `!` to filter the file runs into problems if the formatter errors
out, leaving the file in an unknown or inconsistent state.

Signed-off-by: Christopher Larson <kergoth@gmail.com>