Ignore trailing whitespace on its own line
It's common in a number of cases to keep indentation whitespace on its
own line to indicate paragraph or block continuation. I don't really
care about it that much, and in most cases where I do, I automate
removal of them anyway, so there's no real need to highlight them.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>